requirements.txt
flask
PyMuPDF
python-pptx
Pillow
openai
httpx[socks]
rapidocr-onnxruntime
onnxruntime==1.18.1
